Fix unit tests under new Cypress
