+++ .gitignore
@@ -32,3 +32,4 @@ npm-debug.log*
 
 # misc
 /backup
+.nyc_output
