+++ cypress.json
@@ -2,6 +2,8 @@
   "baseUrl": "http://localhost:3000",
   "viewportWidth": 1920,
   "viewportHeight": 1080,
-  "integrationFolder": "./src",
-  "testFiles": "**/**.test.{ts,tsx}"
+  "integrationFolder": "cypress/integration",
+  "testFiles": "**/**.test.{ts,tsx}",
+  "experimentalComponentTesting": true,
+  "componentFolder": "src"
 }
