+++ package.json
@@ -12,8 +12,8 @@
     "analyze": "source-map-explorer 'build/static/js/*.js'",
     "test": "cypress run --config video=false",
     "test:i": "cypress open",
-    "test:unit": "npm run test -- --spec \"src/**/**.unit.test.tsx\"",
-    "test:e2e": "percy exec -- npm run test -- --spec \"src/**/**.e2e.test.ts\"",
+    "test:unit": "npm run test -- --spec \"./**/**.unit.test.tsx\"",
+    "test:e2e": "percy exec -- npm run test -- --spec \"./**/**.e2e.test.ts\"",
     "test:unit:ci": "npm run start:ci && npm run test:unit",
     "test:e2e:ci": "npm run start:ci && npm run test:e2e",
     "api:generate-types": "apollo codegen:generate types --includes=\"src/**/*.{ts,tsx}\" --endpoint $(grep REACT_APP_DB_GRAPHQL_API_URL .env | cut -d '=' -f2) --target typescript --tagName gql --passthroughCustomScalars --customScalarsPrefix InternPlus && rm -rf types/ && grep -rl InternPlusDateTime ./src | xargs sed -i '' 's/InternPlusDateTime/string/g'"
@@ -82,7 +82,8 @@
     "@typescript-eslint/parser": "^2.12.0",
     "chalk": "^2.4.2",
     "cross-fetch": "^3.0.4",
-    "cypress": "^3.8.0",
+    "cypress": "^4.7.0",
+    "cypress-react-unit-test": "^4.4.2",
     "dotenv": "^8.2.0",
     "eslint-config-prettier": "^6.7.0",
     "eslint-plugin-cypress": "^2.8.1",
