+++ src/components/Button/button.unit.test.tsx
@@ -1,11 +1,12 @@
 import React from "react";
 import { withTheme } from "../../theme/withTheme";
+import { mount } from "cypress-react-unit-test";
 
 import Button from "./index";
 
 describe("Button component", () => {
-  it("works", () => {
-    cy.mount(withTheme(<Button color="greenDark">Sup</Button>));
+  it("renders properly", () => {
+    mount(withTheme(<Button color="greenDark">Sup</Button>));
     cy.get("button").contains("Sup");
   });
 });
