+++ src/pages/reviews/index.tsx
@@ -6,7 +6,7 @@ import { RouteName } from "src/shared/constants/routing";
 import { SearchParamKey, SearchType } from "src/shared/constants/search";
 
 import ReviewPage from "src/pages/reviews/components/ReviewPage";
-import AddReviewModal from "src/pages/reviews/components/AddReviewModalNew";
+import AddReviewModal from "src/pages/reviews/components/AddReviewModal";
 
 /**
  * Page router will either:
